Addition of example files
Examples for various codes that are representative of different situations are presented. The ReadMe contains some comments on what remains to be complete: due to the modification of how the functions were implemented (they were moved out of the folders and into their own folder, who's path is then added in the main scripts of each run).